requirements.txt
requests==2.31.0
langchain==0.0.295
python-dotenv==1.0.0
pathlib
ibm-cloud-sdk-core==3.16.6
ibm-watson==7.0.0
ibm-generative-ai==0.2.8
numpy==1.21.6
toml==0.10.2
ibm-watson-machine-learning==1.0.312
huggingface_hub
unstructured
pypdf
pdfminer
pdf2image
pdfminer-six
streamlit
streamlit-chat
lark
hanziconv==0.3.2
cn2an
selenium
pypandoc==1.11
pymilvus